Restore capabilities on shadow-utils binaries
As of this writing, the registry.fedoraproject.org/fedora:latest image
seems to be missing the security.capability attributes on the
/usr/bin/newuidmap and /usr/bin/newgidmap binaries.  Without them, we
can't properly set up user namespaces when we're run with a UID != 0.

The centos:7 image doesn't include the binaries, and the attributes are
set correctly when we install the shadow-utils package using yum, so we
don't need to make changes to it.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>